refactor] 인수테스트 공통 모듈 추출
